% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{countyDat}
\alias{countyDat}
\title{Data of California Counties by Census Tract}
\format{
A data frame with  1,095,480 rows and 5 columns:
\describe{
  \item{United Statues}{California}
  \item{data downloaded using the get_acs fuction from the tidycensus package}{get_acs}
  \item{2016}{2020}
}
}
\source{
U.S. Census
}
\usage{
countyDat
}
\description{
The data is from the ACS 2016-2020 year
}
\keyword{datasets}
