% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{countyDatMap}
\alias{countyDatMap}
\title{Data of California Counties by Census Tract}
\format{
A data frame with 18,262  rows and 6 columns:
\describe{
  \item{GEOID}{A number composed of State Code, first two digits, then County Code, next two digits, then Census Tract number}
  \item{NAME}{Desctiption of the Tract number}
  \item{variable}{Varible name code}
  \item{estimate}{Estimate of the variable}
  \item{moe}{Margin of Error of variable}
  \item{geometry}{geometry of the Census Tract}
}
}
\source{
U.S. Census
}
\usage{
countyDatMap
}
\description{
The data is from the ACS 2016-2020 year
}
\keyword{datasets}
