% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCensusData.R, R/getMapData.R
\name{getACSData}
\alias{getACSData}
\title{get data from ACS using tidycensus functions}
\usage{
getACSData()

getACSData()
}
\description{
get data from ACS using tidycensus functions

get data from ACS using tidycensus functions
}
\details{
get data from ACS using tidycensus functions

get data from ACS using tidycensus functions
}
\examples{
\dontrun{
getACSData()
 }
\dontrun{
getACSData()
 }
}
