% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wideData.R
\name{wideData}
\alias{wideData}
\title{modified the data to be in wide format for further analysis}
\usage{
wideData(x)
}
\arguments{
\item{x}{a value of the data in long format}
}
\description{
modified the data to be in wide format for further analysis
}
\examples{
\dontrun{
wideData(x)
 }
}
